rename: entries->pages
