+++ frontend/src/layouts/Main.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import { Switch, Route } from "react-router-dom";
 
-import Index from "@/entries/Index";
-import RadioHistory from "@/entries/RadioHistory";
-import Contact from "@/entries/Contact";
-import Blog from "@/entries/Blog";
-import Personality from "@/entries/Personality";
+import Index from "@/pages/Index";
+import RadioHistory from "@/pages/RadioHistory";
+import Contact from "@/pages/Contact";
+import Blog from "@/pages/Blog";
+import Personality from "@/pages/Personality";
 
 export default () => (
   <main>
