+++ frontend/src/pages/Personality.tsx
